requirements.txt
Pillow==9.2.0
streamlit==1.12.0
altair==4.2.0